create purchase logic, fixed and changed some exceptions

+++ src/main/java/itAcademy/giraffeMT/giraffeMT/services/impl/UserServiceImpl.java
@@ -4,7 +4,7 @@ import com.company.banksystem.entity.BankAccount;
 import com.company.banksystem.service.interfaces.BankAccountService;
 import itAcademy.giraffeMT.giraffeMT.entities.User;
 import itAcademy.giraffeMT.giraffeMT.exceptions.NotFound;
-import itAcademy.giraffeMT.giraffeMT.models.UserModel;
+import itAcademy.giraffeMT.giraffeMT.dto.UserModel;
 import itAcademy.giraffeMT.giraffeMT.repositories.UserRepository;
 import itAcademy.giraffeMT.giraffeMT.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,8 +17,8 @@ import java.util.Optional;
 public class UserServiceImpl implements UserService {
     @Autowired
     private UserRepository userRepository;
-    //  @Autowired
-    // private BankAccountService bankAccountService;
+
+    @Autowired
     private BankAccountService bankAccountService;
 
     @Override
@@ -34,14 +34,22 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User create(UserModel model) throws NotFound {
-//        if (model.getRequisite() != null) {
+        if (model.getRequisite() != null) {
             BankAccount bankAccount = bankAccountService.findBankAccountByAccountNumber(model.getRequisite());
+            if (bankAccount != null) {
+                User user = User.builder().fullName(model.getFullName())
+                        .login(model.getLogin())
+                        .password(model.getPassword())
+                        .bankAccount(bankAccount).build();
+                return userRepository.save(user);
+            } else throw new NotFound("bank account doesn't exist");
+        } else {
             User user = User.builder().fullName(model.getFullName())
                     .login(model.getLogin())
                     .password(model.getPassword())
-                    .bankAccount(bankAccount).build();
+                    .build();
             return userRepository.save(user);
-         // throw new NotFound("bank account not found");
+        }
     }
 
     @Override
@@ -55,13 +63,18 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void delete(Long id) throws Exception {
-
+        User user = getById(id);
+        if (user != null)
+            userRepository.delete(user);
+        throw new NotFound("User doesn't exist");
     }
 
     @Override
     public void deleteByLogin(String login) throws NotFound {
         User user = findByLogin(login);
-        userRepository.delete(user);
+        if (user != null)
+            userRepository.delete(user);
+        throw new NotFound("User doesn't exist");
     }
 
     @Override
